refactor: rename action and atom file name
