+++ src/ts/index.tsx
@@ -19,7 +19,7 @@ import { EventDispatcherImpl } from "./infrastractures/event/event-dispatcher";
 import { GameCreatedEventListener } from "./infrastractures/event/game-created-event-listener";
 import { GameRepositoryImpl } from "./infrastractures/game-repository";
 import { GameShowedDownEventListener } from "./infrastractures/event/game-showed-down-event-listener";
-import { createInGameAction } from "./status/in-game-action";
+import { createInGameAction } from "./status/game-action";
 import { HandCardUseCase } from "./usecases/hand-card";
 import { ShowDownUseCase } from "./usecases/show-down";
 import { NewGameUseCase } from "./usecases/new-game";
@@ -33,7 +33,7 @@ import { ChangeUserNameUseCase } from "./usecases/change-user-name";
 import { ChangeUserModeUseCase } from "./usecases/change-user-mode";
 import { GamePlayerRepositoryImpl } from "./infrastractures/game-player-repository";
 import { createJoinService } from "./domains/join-service";
-import { createInGameSelectors } from "./status/in-game-selector";
+import { createInGameSelectors } from "./status/game-selector";
 import { UserLeaveFromGameEventListener } from "./infrastractures/event/user-leave-from-game-event-listener";
 import { createDependencyRegistrar } from "./utils/dependency-registrar";
 import { ApplicationDependencyRegistrar } from "./dependencies";
